Rename harness to compiler-io for clarity
